+++ src/helpers.ts
@@ -225,14 +225,13 @@ Handlebars.registerHelper('useAxis', function (
 Handlebars.registerHelper('setupAxis', function (
     id: string,
     method: string,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    ...args: any[]
+    ...args: unknown[]
 ) {
     const axis = axes.get(id);
     if (axis) {
         args.pop();
         if (method === 'tickFormat') {
-            axis.tickFormat(format(args[0]));
+            axis.tickFormat(format(args[0] as string));
         } else {
             axis[method].call(axis, ...args);
         }
@@ -242,8 +241,7 @@ Handlebars.registerHelper('setupAxis', function (
 Handlebars.registerHelper('getScale', function (
     id: string,
     method: string,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    ...args: any[]
+    ...args: unknown[]
 ) {
     const scale = scales.get(id);
     if (scale) {
@@ -255,8 +253,7 @@ Handlebars.registerHelper('getScale', function (
 Handlebars.registerHelper('setupScale', function (
     id: string,
     method: string,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    ...args: any[]
+    ...args: unknown[]
 ) {
     const scale = scales.get(id);
     if (scale) {
@@ -308,8 +305,7 @@ Handlebars.registerHelper('val', function (
 
 Handlebars.registerHelper('math', function (
     method: string,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    ...args: any[]
+    ...args: unknown[]
 ) {
     args.pop()
     if (args.length) {