+++ src/redux/slice.ts
@@ -4,6 +4,7 @@ import IVisualHost = powerbiVisualsApi.extensibility.visual.IVisualHost;
 import IViewport = powerbiVisualsApi.IViewport;
 import ISelectionManager = powerbiVisualsApi.extensibility.ISelectionManager;
 import ISelectionId = powerbiVisualsApi.visuals.ISelectionId;
+import ILocalizationManager = powerbiVisualsApi.extensibility.ILocalizationManager;
 
 import { createSlice } from '@reduxjs/toolkit';
 import type { PayloadAction } from '@reduxjs/toolkit';
@@ -20,6 +21,8 @@ export interface VisualState {
     selectionManager: ISelectionManager | undefined;
     selectionIds: ISelectionId[];
     tooltipColumns: TooltipColumnData[];
+    isHighContrast: boolean;
+    localizationManager: ILocalizationManager | undefined;
 }
 
 export interface RowDataPayload {
@@ -40,7 +43,9 @@ const initialState: VisualState = {
     markdownSections: [],
     selectionManager: undefined,
     selectionIds: [],
-    tooltipColumns: []
+    tooltipColumns: [],
+    isHighContrast: false,
+    localizationManager: undefined,
 }
 
 export const slice = createSlice({
@@ -69,11 +74,17 @@ export const slice = createSlice({
         },
         setSettings: (state, action: PayloadAction<IVisualSettings>) => {
             state.settings = action.payload;
-        }
+        },
+        setHighContrast: (state, action: PayloadAction<boolean>) => {
+            state.isHighContrast = action.payload;
+        },
+        setLocalizationManager: (state, action: PayloadAction<ILocalizationManager>) => {
+            state.localizationManager = action.payload;
+        },
     },
 })
 
 // Action creators are generated for each case reducer function
-export const { setHost, setSelectionManager, setDataView, setRowData, setSettings, setViewport } = slice.actions
+export const { setHost, setSelectionManager, setDataView, setRowData, setSettings, setViewport, setHighContrast, setLocalizationManager } = slice.actions
 
 export default slice.reducer